pyproject.toml
[project]
name = "mgsplus"
version = "0.1.0"
description = "MGSPlus multi-agent system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Existing
    "kaggle>=1.7.4.5",
    "kagglehub>=0.3.13",
    "pyprojroot>=0.3.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    # Agent framework
    "crewai[litellm]>=1.0.0",
    "crewai-tools>=1.0.0",
    "litellm>=1.74.9",
    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    # Databases
    "qdrant-client>=1.9.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "pymssql>=2.3.0",
    "neo4j>=5.20.0",
    # MCP
    "mcp>=1.0.0",
    # Config / schemas
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # Async
    "anyio>=4.4.0",
    "aioodbc>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[project.scripts]
agents = "src.agents.server:run"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
testpaths = ["src/agents/tests"]
